refactor(wallet): extract useSolPrice, useDisplayPreferences, useTelegramSetup hooks
